chore(ex00): modify makefile

+++ ex00/megaphone.cpp
@@ -22,7 +22,7 @@ int	main(int argc, char **argv)
 	{
 		for (int i = 1; i < argc; i++)
 		{
-			for (int j = 0; j < std::strlen(argv[i]); j++)
+			for (size_t j = 0; j < std::strlen(argv[i]); j++)
 				std::cout << (char)std::toupper(argv[i][j]);
 		}
 	}
